==> REST[ES]_donors/Donor10/2_Normalize_for_donor.Rmd <==
---
title: "Normalize Data"
author: "Wesley Burr"
date: "29/03/2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(dplyr)
```

## Normalize Data

We will load each of the data files, identify the 'primary' peak for
Bromobenze, then normalize all Areas to be relative areas to the
reference standard.

```{r}
load("/home/wburr/Rushali_REST/Donor10/D10.rda")
D10m$Area<-as.numeric(D10m$Area)
bromo <- D10m %>% 
           group_by(Sample) %>% 
           subset(Name == "Benzene, bromo-") %>%
           filter(Area == max(Area))
split_Donor <- D10m %>% 
                group_by(Sample) %>%
                group_split()
for(j in 1:length(split_Donor)) {
  split_Donor[[j]]$Area <- split_Donor[[j]]$Area / unlist(bromo[j, "Area"])
}
D10m <- do.call("rbind", split_Donor)
save(file = "/home/wburr/Rushali_REST/Donor10/D10n.rda", D10m)
```
